refactor: decompose Topbar, split UI store, consolidate auth, route-aware shell
Architecture improvements based on code review:

1. **Store slices** — Split monolithic useUIStore into focused slices
   (layout, toast, lens, status) composed via Zustand's slice pattern.
   Same API surface, zero consumer changes needed. Removed dead state
   (guideOpen, firstWinDismissed).

2. **Topbar decomposition** — Extracted LensTitle, NotificationBell,
   and UserMenu into focused sub-components under components/shell/topbar/.
   Each manages its own state and event listeners instead of one god component.

3. **Route-aware AppShell** — AppShell now checks pathname against
   STANDALONE_PREFIXES (e.g. /legal/) to skip shell chrome, removing
   the need for pages to manually toggle fullPageMode via useEffect.
   Legal layout simplified (no more store dependency).

4. **Auth cookie consolidation** — Middleware now only checks
   concord_auth and concord_refresh (the cookies actually set by the
   server). Removed legacy cookie names (token, connect.sid,
   concord_session) that were never set in production.

5. **API interceptor fix** — 401 handler now skips redirect on public
   pages (/legal/*, /, /login, /register, /forgot-password) to prevent
   incidental API calls from triggering login redirects.

All 1959 unit tests pass. All e2e auth helpers updated to use
concord_auth. TypeScript and lint clean.
