+++ concord-frontend/store/slices/layout.ts
@@ -1,0 +1,33 @@
+import type { StateCreator } from 'zustand';
+
+export interface LayoutSlice {
+  sidebarOpen: boolean;
+  sidebarCollapsed: boolean;
+  commandPaletteOpen: boolean;
+  theme: 'dark' | 'light';
+  fullPageMode: boolean;
+
+  toggleSidebar: () => void;
+  setSidebarOpen: (open: boolean) => void;
+  setSidebarCollapsed: (collapsed: boolean) => void;
+  toggleCommandPalette: () => void;
+  setCommandPaletteOpen: (open: boolean) => void;
+  setTheme: (theme: 'dark' | 'light') => void;
+  setFullPageMode: (mode: boolean) => void;
+}
+
+export const createLayoutSlice: StateCreator<LayoutSlice, [], [], LayoutSlice> = (set) => ({
+  sidebarOpen: true,
+  sidebarCollapsed: false,
+  commandPaletteOpen: false,
+  theme: 'dark',
+  fullPageMode: false,
+
+  toggleSidebar: () => set((state) => ({ sidebarOpen: !state.sidebarOpen })),
+  setSidebarOpen: (open) => set({ sidebarOpen: open }),
+  setSidebarCollapsed: (collapsed) => set({ sidebarCollapsed: collapsed }),
+  toggleCommandPalette: () => set((state) => ({ commandPaletteOpen: !state.commandPaletteOpen })),
+  setCommandPaletteOpen: (open) => set({ commandPaletteOpen: open }),
+  setTheme: (theme) => set({ theme }),
+  setFullPageMode: (mode) => set({ fullPageMode: mode }),
+});
