+++ concord-frontend/store/slices/toast.ts
@@ -1,0 +1,31 @@
+import type { StateCreator } from 'zustand';
+
+export interface Toast {
+  id: string;
+  type: 'success' | 'error' | 'warning' | 'info';
+  message: string;
+  duration?: number;
+}
+
+export interface ToastSlice {
+  toasts: Toast[];
+  addToast: (toast: Omit<Toast, 'id'>) => void;
+  removeToast: (id: string) => void;
+}
+
+export const createToastSlice: StateCreator<ToastSlice, [], [], ToastSlice> = (set) => ({
+  toasts: [],
+
+  addToast: (toast) =>
+    set((state) => ({
+      toasts: [
+        ...state.toasts,
+        { ...toast, id: `toast-${Date.now()}-${Math.random().toString(36).substr(2, 9)}` },
+      ],
+    })),
+
+  removeToast: (id) =>
+    set((state) => ({
+      toasts: state.toasts.filter((t) => t.id !== id),
+    })),
+});
